+++ tests/back/Pim/WorkOrganization/Specification/Workflow/Bundle/Presenter/BooleanPresenterSpec.php
@@ -9,6 +9,14 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 
 class BooleanPresenterSpec extends ObjectBehavior
 {
+    function let(
+        RendererInterface $renderer,
+        TranslatorInterface $translator
+    ) {
+        $this->setTranslator($translator);
+        $this->setRenderer($renderer);
+    }
+
     function it_is_a_translator_aware_presenter()
     {
         $this->shouldBeAnInstanceOf(PresenterInterface::class);
@@ -24,13 +32,23 @@ class BooleanPresenterSpec extends ObjectBehavior
         RendererInterface $renderer,
         TranslatorInterface $translator
     ) {
-        $translator->trans('Yes')->willReturn('Yes');
-        $translator->trans('No')->willReturn('No');
+        $translator->trans('Yes')->shouldBeCalled()->willReturn('Yes');
+        $translator->trans('No')->shouldBeCalled()->willReturn('No');
+        $renderer->renderDiff('No', 'Yes')->shouldBeCalled()->willReturn('diff between two booleans');
 
-        $renderer->renderDiff('No', 'Yes')->willReturn('diff between two booleans');
+        $this->present(false, ['data' => '1'])->shouldReturn('diff between two booleans');
+    }
 
-        $this->setRenderer($renderer);
-        $this->setTranslator($translator);
-        $this->present(['data' => false], ['data' => '1'])->shouldReturn('diff between two booleans');
+    function it_presents_an_empty_boolean_value_as_an_empty_string(
+        RendererInterface $renderer,
+        TranslatorInterface $translator
+    ) {
+        $translator->trans('Yes')->shouldBeCalled()->willReturn('Yes');
+        $renderer->renderDiff('', 'Yes')->shouldBeCalled()->willReturn('diff between empty and boolean');
+        $this->present(null, ['data' => true])->shouldReturn('diff between empty and boolean');
+
+        $translator->trans('No')->shouldBeCalled()->willReturn('No');
+        $renderer->renderDiff('No', '')->shouldBeCalled()->willReturn('diff between boolean and empty');
+        $this->present(false, ['data' => null])->shouldReturn('diff between boolean and empty');
     }
 }